Create gg.py
I am very happy to announce gg which means gpaw-tools GUI. It is the GUI for gpawsolve.py for now and it is brewing for the last few weeks separate from the main project.  It can run all gpawsolve.py's features with a tk GUI. This first version does not include ciftoase.py or optimization codes features, however, these wished features can be accepted as a to-do.